- [ADDITION] Added abstract base fragment class (FragmentDialog)
